+++ plugins/modules/oci_data_science_model_version_set.py
@@ -1,0 +1,444 @@
+#!/usr/bin/python
+# Copyright (c) 2020, 2023 Oracle and/or its affiliates.
+# This software is made available to you under the terms of the GPL 3.0 license or the Apache 2.0 license.
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+# Apache License v2.0
+# See LICENSE.TXT for details.
+# GENERATED FILE - DO NOT EDIT - MANUAL CHANGES WILL BE OVERWRITTEN
+
+
+from __future__ import absolute_import, division, print_function
+
+__metaclass__ = type
+
+ANSIBLE_METADATA = {
+    "metadata_version": "1.1",
+    "status": ["preview"],
+    "supported_by": "community",
+}
+
+DOCUMENTATION = """
+---
+module: oci_data_science_model_version_set
+short_description: Manage a ModelVersionSet resource in Oracle Cloud Infrastructure
+description:
+    - This module allows the user to create, update and delete a ModelVersionSet resource in Oracle Cloud Infrastructure
+    - For I(state=present), creates a new modelVersionSet.
+    - "This resource has the following action operations in the M(oracle.oci.oci_data_science_model_version_set_actions) module: change_compartment."
+version_added: "2.9.0"
+author: Oracle (@oracle)
+options:
+    compartment_id:
+        description:
+            - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the compartment to create the model version set in.
+            - Required for create using I(state=present).
+            - Required for update when environment variable C(OCI_USE_NAME_AS_IDENTIFIER) is set.
+            - Required for delete when environment variable C(OCI_USE_NAME_AS_IDENTIFIER) is set.
+        type: str
+    project_id:
+        description:
+            - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the project to associate with the model version set.
+            - Required for create using I(state=present).
+        type: str
+    name:
+        description:
+            - "A user-friendly name for the resource. It must be unique and can't be modified. Avoid entering confidential information.
+              Example: `My model version set`"
+            - Required for create using I(state=present).
+            - Required for update, delete when environment variable C(OCI_USE_NAME_AS_IDENTIFIER) is set.
+        type: str
+    description:
+        description:
+            - A short description of the model version set.
+            - This parameter is updatable.
+        type: str
+    freeform_tags:
+        description:
+            - "Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. See L(Resource
+              Tags,https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
+              Example: `{\\"Department\\": \\"Finance\\"}`"
+            - This parameter is updatable.
+        type: dict
+    defined_tags:
+        description:
+            - "Defined tags for this resource. Each key is predefined and scoped to a namespace. See L(Resource
+              Tags,https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
+              Example: `{\\"Operations\\": {\\"CostCenter\\": \\"42\\"}}`"
+            - This parameter is updatable.
+        type: dict
+    model_version_set_id:
+        description:
+            - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the model version set.
+            - Required for update using I(state=present) when environment variable C(OCI_USE_NAME_AS_IDENTIFIER) is not set.
+            - Required for delete using I(state=absent) when environment variable C(OCI_USE_NAME_AS_IDENTIFIER) is not set.
+        type: str
+        aliases: ["id"]
+    is_delete_related_models:
+        description:
+            - By default, this parameter is false. A model version set can only be deleted if all the models associate with it are already in the DELETED state.
+              You can optionally specify the deleteRelatedModels boolean query parameters to true, which deletes all associated models for you.
+        type: bool
+    state:
+        description:
+            - The state of the ModelVersionSet.
+            - Use I(state=present) to create or update a ModelVersionSet.
+            - Use I(state=absent) to delete a ModelVersionSet.
+        type: str
+        required: false
+        default: 'present'
+        choices: ["present", "absent"]
+extends_documentation_fragment: [ oracle.oci.oracle, oracle.oci.oracle_creatable_resource, oracle.oci.oracle_wait_options ]
+"""
+
+EXAMPLES = """
+- name: Create model_version_set
+  oci_data_science_model_version_set:
+    # required
+    compartment_id: "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx"
+    project_id: "ocid1.project.oc1..xxxxxxEXAMPLExxxxxx"
+    name: name_example
+
+    # optional
+    description: description_example
+    freeform_tags: {'Department': 'Finance'}
+    defined_tags: {'Operations': {'CostCenter': 'US'}}
+
+- name: Update model_version_set
+  oci_data_science_model_version_set:
+    # required
+    model_version_set_id: "ocid1.modelversionset.oc1..xxxxxxEXAMPLExxxxxx"
+
+    # optional
+    description: description_example
+    freeform_tags: {'Department': 'Finance'}
+    defined_tags: {'Operations': {'CostCenter': 'US'}}
+
+- name: Update model_version_set using name (when environment variable OCI_USE_NAME_AS_IDENTIFIER is set)
+  oci_data_science_model_version_set:
+    # required
+    compartment_id: "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx"
+    name: name_example
+
+    # optional
+    description: description_example
+    freeform_tags: {'Department': 'Finance'}
+    defined_tags: {'Operations': {'CostCenter': 'US'}}
+
+- name: Delete model_version_set
+  oci_data_science_model_version_set:
+    # required
+    model_version_set_id: "ocid1.modelversionset.oc1..xxxxxxEXAMPLExxxxxx"
+    state: absent
+
+    # optional
+    is_delete_related_models: true
+
+- name: Delete model_version_set using name (when environment variable OCI_USE_NAME_AS_IDENTIFIER is set)
+  oci_data_science_model_version_set:
+    # required
+    compartment_id: "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx"
+    name: name_example
+    state: absent
+
+"""
+
+RETURN = """
+model_version_set:
+    description:
+        - Details of the ModelVersionSet resource acted upon by the current operation
+    returned: on success
+    type: complex
+    contains:
+        id:
+            description:
+                - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the model version set.
+            returned: on success
+            type: str
+            sample: "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx"
+        compartment_id:
+            description:
+                - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the model version set compartment.
+            returned: on success
+            type: str
+            sample: "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx"
+        project_id:
+            description:
+                - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the project associated with the model version set.
+            returned: on success
+            type: str
+            sample: "ocid1.project.oc1..xxxxxxEXAMPLExxxxxx"
+        name:
+            description:
+                - A user-friendly name for the resource.
+            returned: on success
+            type: str
+            sample: name_example
+        description:
+            description:
+                - A short description of the model version set.
+            returned: on success
+            type: str
+            sample: description_example
+        lifecycle_state:
+            description:
+                - The state of the model version set.
+            returned: on success
+            type: str
+            sample: ACTIVE
+        time_created:
+            description:
+                - "The date and time that the resource was created in the timestamp format defined by L(RFC3339,https://tools.ietf.org/html/rfc3339).
+                  Example: 2019-08-25T21:10:29.41Z"
+            returned: on success
+            type: str
+            sample: "2013-10-20T19:20:30+01:00"
+        time_updated:
+            description:
+                - "The date and time that the resource was created in the timestamp format defined by L(RFC3339,https://tools.ietf.org/html/rfc3339).
+                  Example: 2019-08-25T21:10:29.41Z"
+            returned: on success
+            type: str
+            sample: "2013-10-20T19:20:30+01:00"
+        created_by:
+            description:
+                - The L(OCID,https://docs.cloud.oracle.com/iaas/Content/General/Concepts/identifiers.htm) of the user who created the model version set.
+            returned: on success
+            type: str
+            sample: created_by_example
+        freeform_tags:
+            description:
+                - "Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. See L(Resource
+                  Tags,https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
+                  Example: `{\\"Department\\": \\"Finance\\"}`"
+            returned: on success
+            type: dict
+            sample: {'Department': 'Finance'}
+        defined_tags:
+            description:
+                - "Defined tags for this resource. Each key is predefined and scoped to a namespace. See L(Resource
+                  Tags,https://docs.cloud.oracle.com/Content/General/Concepts/resourcetags.htm).
+                  Example: `{\\"Operations\\": {\\"CostCenter\\": \\"42\\"}}`"
+            returned: on success
+            type: dict
+            sample: {'Operations': {'CostCenter': 'US'}}
+    sample: {
+        "id": "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx",
+        "compartment_id": "ocid1.compartment.oc1..xxxxxxEXAMPLExxxxxx",
+        "project_id": "ocid1.project.oc1..xxxxxxEXAMPLExxxxxx",
+        "name": "name_example",
+        "description": "description_example",
+        "lifecycle_state": "ACTIVE",
+        "time_created": "2013-10-20T19:20:30+01:00",
+        "time_updated": "2013-10-20T19:20:30+01:00",
+        "created_by": "created_by_example",
+        "freeform_tags": {'Department': 'Finance'},
+        "defined_tags": {'Operations': {'CostCenter': 'US'}}
+    }
+"""
+
+from ansible_collections.oracle.oci.plugins.module_utils import (
+    oci_common_utils,
+    oci_wait_utils,
+)
+from ansible_collections.oracle.oci.plugins.module_utils.oci_resource_utils import (
+    OCIResourceHelperBase,
+    get_custom_class,
+    OCIAnsibleModule,
+)
+
+try:
+    from oci.data_science import DataScienceClient
+    from oci.data_science.models import CreateModelVersionSetDetails
+    from oci.data_science.models import UpdateModelVersionSetDetails
+
+    HAS_OCI_PY_SDK = True
+except ImportError:
+    HAS_OCI_PY_SDK = False
+
+
+class DataScienceModelVersionSetHelperGen(OCIResourceHelperBase):
+    """Supported operations: create, update, get, list and delete"""
+
+    def get_possible_entity_types(self):
+        return super(
+            DataScienceModelVersionSetHelperGen, self
+        ).get_possible_entity_types() + [
+            "modelversionset",
+            "modelversionsets",
+            "dataSciencemodelversionset",
+            "dataSciencemodelversionsets",
+            "modelversionsetresource",
+            "modelversionsetsresource",
+            "datascience",
+        ]
+
+    def get_module_resource_id_param(self):
+        return "model_version_set_id"
+
+    def get_module_resource_id(self):
+        return self.module.params.get("model_version_set_id")
+
+    def get_get_fn(self):
+        return self.client.get_model_version_set
+
+    def get_get_model_from_summary_model(self, summary_model):
+        return oci_common_utils.call_with_backoff(
+            self.client.get_model_version_set, model_version_set_id=summary_model.id,
+        ).data
+
+    def get_resource(self):
+        return oci_common_utils.call_with_backoff(
+            self.client.get_model_version_set,
+            model_version_set_id=self.module.params.get("model_version_set_id"),
+        )
+
+    def get_required_kwargs_for_list(self):
+        required_list_method_params = [
+            "compartment_id",
+        ]
+
+        return dict(
+            (param, self.module.params[param]) for param in required_list_method_params
+        )
+
+    def get_optional_kwargs_for_list(self):
+        optional_list_method_params = ["project_id", "name"]
+
+        return dict(
+            (param, self.module.params[param])
+            for param in optional_list_method_params
+            if self.module.params.get(param) is not None
+            and (
+                self._use_name_as_identifier()
+                or (
+                    not self.module.params.get("key_by")
+                    or param in self.module.params.get("key_by")
+                )
+            )
+        )
+
+    def list_resources(self):
+
+        required_kwargs = self.get_required_kwargs_for_list()
+        optional_kwargs = self.get_optional_kwargs_for_list()
+        kwargs = oci_common_utils.merge_dicts(required_kwargs, optional_kwargs)
+        return oci_common_utils.list_all_resources(
+            self.client.list_model_version_sets, **kwargs
+        )
+
+    def get_create_model_class(self):
+        return CreateModelVersionSetDetails
+
+    def create_resource(self):
+        create_details = self.get_create_model()
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.create_model_version_set,
+            call_fn_args=(),
+            call_fn_kwargs=dict(create_model_version_set_details=create_details,),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.CREATE_OPERATION_KEY,
+            waiter_client=self.get_waiter_client(),
+            resource_helper=self,
+            wait_for_states=self.get_wait_for_states_for_operation(
+                oci_common_utils.CREATE_OPERATION_KEY,
+            ),
+        )
+
+    def get_update_model_class(self):
+        return UpdateModelVersionSetDetails
+
+    def update_resource(self):
+        update_details = self.get_update_model()
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.update_model_version_set,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                model_version_set_id=self.module.params.get("model_version_set_id"),
+                update_model_version_set_details=update_details,
+            ),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.UPDATE_OPERATION_KEY,
+            waiter_client=self.get_waiter_client(),
+            resource_helper=self,
+            wait_for_states=self.get_wait_for_states_for_operation(
+                oci_common_utils.UPDATE_OPERATION_KEY,
+            ),
+        )
+
+    def delete_resource(self):
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.delete_model_version_set,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                model_version_set_id=self.module.params.get("model_version_set_id"),
+                is_delete_related_models=self.module.params.get(
+                    "is_delete_related_models"
+                ),
+            ),
+            waiter_type=oci_wait_utils.WORK_REQUEST_WAITER_KEY,
+            operation=oci_common_utils.DELETE_OPERATION_KEY,
+            waiter_client=self.get_waiter_client(),
+            resource_helper=self,
+            wait_for_states=oci_common_utils.get_work_request_completed_states(),
+        )
+
+
+DataScienceModelVersionSetHelperCustom = get_custom_class(
+    "DataScienceModelVersionSetHelperCustom"
+)
+
+
+class ResourceHelper(
+    DataScienceModelVersionSetHelperCustom, DataScienceModelVersionSetHelperGen
+):
+    pass
+
+
+def main():
+    module_args = oci_common_utils.get_common_arg_spec(
+        supports_create=True, supports_wait=True
+    )
+    module_args.update(
+        dict(
+            compartment_id=dict(type="str"),
+            project_id=dict(type="str"),
+            name=dict(type="str"),
+            description=dict(type="str"),
+            freeform_tags=dict(type="dict"),
+            defined_tags=dict(type="dict"),
+            model_version_set_id=dict(aliases=["id"], type="str"),
+            is_delete_related_models=dict(type="bool"),
+            state=dict(type="str", default="present", choices=["present", "absent"]),
+        )
+    )
+
+    module = OCIAnsibleModule(argument_spec=module_args, supports_check_mode=True)
+
+    if not HAS_OCI_PY_SDK:
+        module.fail_json(msg="oci python sdk required for this module.")
+
+    resource_helper = ResourceHelper(
+        module=module,
+        resource_type="model_version_set",
+        service_client_class=DataScienceClient,
+        namespace="data_science",
+    )
+
+    result = dict(changed=False)
+
+    if resource_helper.is_delete_using_name():
+        result = resource_helper.delete_using_name()
+    elif resource_helper.is_delete():
+        result = resource_helper.delete()
+    elif resource_helper.is_update_using_name():
+        result = resource_helper.update_using_name()
+    elif resource_helper.is_update():
+        result = resource_helper.update()
+    elif resource_helper.is_create():
+        result = resource_helper.create()
+
+    module.exit_json(**result)
+
+
+if __name__ == "__main__":
+    main()