+++ plugins/modules/oci_service_mesh_ingress_gateway_route_table.py
@@ -154,6 +154,15 @@ options:
                     - If true, the matched path prefix will be rewritten to '/' before being directed to the target virtual deployment.
                     - Applicable when type is 'HTTP'
                 type: bool
+            request_timeout_in_ms:
+                description:
+                    - The maximum duration in milliseconds for the upstream service to respond to a request.
+                      If provided, the timeout value overrides the default timeout of 15 seconds for the HTTP based route rules, and disabled (no timeout) when
+                      'isGrpc' is true.
+                      The value 0 (zero) indicates that the timeout is disabled.
+                      For streaming responses from the upstream service, consider either keeping the timeout disabled or set a sufficiently high value.
+                    - Applicable when type is 'HTTP'
+                type: int
     freeform_tags:
         description:
             - "Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
@@ -378,6 +387,16 @@ ingress_gateway_route_table:
                     returned: on success
                     type: bool
                     sample: true
+                request_timeout_in_ms:
+                    description:
+                        - The maximum duration in milliseconds for the upstream service to respond to a request.
+                          If provided, the timeout value overrides the default timeout of 15 seconds for the HTTP based route rules, and disabled (no timeout)
+                          when 'isGrpc' is true.
+                          The value 0 (zero) indicates that the timeout is disabled.
+                          For streaming responses from the upstream service, consider either keeping the timeout disabled or set a sufficiently high value.
+                    returned: on success
+                    type: int
+                    sample: 56
                 type:
                     description:
                         - Type of protocol.
@@ -492,6 +511,7 @@ ingress_gateway_route_table:
             "is_grpc": true,
             "is_host_rewrite_enabled": true,
             "is_path_rewrite_enabled": true,
+            "request_timeout_in_ms": 56,
             "type": "HTTP",
             "ingress_gateway_host": {
                 "name": "name_example",
@@ -720,6 +740,7 @@ def main():
                     is_grpc=dict(type="bool"),
                     is_host_rewrite_enabled=dict(type="bool"),
                     is_path_rewrite_enabled=dict(type="bool"),
+                    request_timeout_in_ms=dict(type="int"),
                 ),
             ),
             freeform_tags=dict(type="dict"),