+++ plugins/modules/oci_service_mesh_virtual_service_route_table_actions.py
@@ -131,6 +131,16 @@ virtual_service_route_table:
                     returned: on success
                     type: bool
                     sample: true
+                request_timeout_in_ms:
+                    description:
+                        - The maximum duration in milliseconds for the target service to respond to a request.
+                          If provided, the timeout value overrides the default timeout of 15 seconds for the HTTP based route rules, and disabled (no timeout)
+                          when 'isGrpc' is true.
+                          The value 0 (zero) indicates that the timeout is disabled.
+                          For streaming responses from the target service, consider either keeping the timeout disabled or set a sufficiently high value.
+                    returned: on success
+                    type: int
+                    sample: 56
                 type:
                     description:
                         - Type of protocol.
@@ -225,6 +235,7 @@ virtual_service_route_table:
             "path": "path_example",
             "path_type": "PREFIX",
             "is_grpc": true,
+            "request_timeout_in_ms": 56,
             "type": "HTTP",
             "destinations": [{
                 "type": "VIRTUAL_DEPLOYMENT",